refactor: adjust date handling and update import paths
Changed `data_nascimento` in `DeveloperDto` to use a string format
instead of a Date type. Added a private helper to format dates to
`YYYY-MM-DD` in the mapper. Also updated incorrect import paths for
better module organization.
