+++ backend/src/application/dtos/developer.dto.ts
@@ -1,7 +1,7 @@
 import { DeveloperRepositoryToServiceDto } from './developerRepositoryToService.dto';
 
 export class DeveloperDto extends DeveloperRepositoryToServiceDto{
-  data_nascimento: Date;
+  data_nascimento: string;
   idade: number;
 
   declare dataNascimento: never;
